dai-zephyr: fix crash in dai_reset
It is not guaranteed that 'dd->z_config' is set when dai_reset()
is called. E.g. if the requested runtime parameters are not
supported by hardare, the parameter setup will not be completed.
When dai_reset() is called in this state, a FW crash results
via NULL dereference.

Fix the issue by checking dd->z_config value before use.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>