zephyr: move running sof_boot ztest suite to platform init
Instead of triggering the boot tests from one of the IPC4 messages,
run the boot test suite already in boot_complete(). This is
more practical to run tests that take a long time to run. If
triggering such tests in middle of IPC sequences, this will lead
to triggering IPC timeouts in SOF Linux driver and other side-effects.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>