[RFC] audio: concept for endpoint-driven ll-scheduler
This patch implements a minimal prototype to separate endpoint
components (host DMA and DAI) from the audio pipelines, and run
the copy() for these endspoints at a separate stage of the low-latency
scheduler.

This change attempts to group all interaction with hardware devices
(DMAs and DAIs) into a single place of audio graph execution.

Notable limitations:
 - the implementation is not complete. this is enough to run real-life
   audio pipelines, but many corner cases are not handled, IOW this
   is just for design discussion
 - clashes with the delayed IPC handling for pipelines
	- pipeline-schedule.c:pipeline_task_cmd() -- these are run
          in middle of pipeline execution and cause problems to endpoint
          handling added in this commit

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>