[RFC] lib: dma: add support for user-space usage
Extend the DMA interface with system call support that allows
to use DMA from user-space.

Use a simple object lookup for kernel objects passed from user-space
and check the object is a valid kernel object before proceeding
to syscall implementation. This is an alternative to Zephyr k_obj_core
infrastructure, which is difficult to use from out-of-tree module
like SOF (e.g. addition of SOF types to gen_kobject_list.py is currently
not feasible out-of-tree).

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>